+++ packages/container/src/tokens.ts
@@ -1,0 +1,3 @@
+export const marginX = `1rem`;
+export const xlMaxWidth = '71.25rem';
+export const responsiveMaxWidth = { sm: '33.75rem', md: '45rem', lg: '60rem', xl: xlMaxWidth };