+++ packages/grid/src/index.doc.tsx
@@ -3,17 +3,13 @@ import styled from 'styled-components';
 import { p } from '@lendi-ui/spacing';
 import { bg, color } from '@lendi-ui/color';
 import { body } from '@lendi-ui/typography';
-import { grid, unit } from '.';
+import Grid from '.';
 
-const Grid = styled.div`
-  ${grid};
-`;
-
-const Unit = styled.code`
+const Content = styled.code`
+  display: block;
   ${p('xxs')}
   ${bg('primary.300')}
   ${body({ size: 'xs', align: 'center' })}
-  ${unit}
   font-weight: bold;
   border: 1px solid ${color('shade.0')}
 `;
@@ -21,33 +17,63 @@ const Unit = styled.code`
 const Story = () => (
   <>
     <Grid>
-      <Unit>1</Unit>
+      <Grid.Unit>
+        <Content>1</Content>
+      </Grid.Unit>
     </Grid>
 
     <Grid>
-      <Unit size={1 / 2}>1/2</Unit>
-      <Unit size={1 / 2}>1/2</Unit>
+      <Grid.Unit size={1 / 2}>
+        <Content>1/2</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 2}>
+        <Content>1/2</Content>
+      </Grid.Unit>
     </Grid>
 
     <Grid>
-      <Unit size={1 / 3}>1/2</Unit>
-      <Unit size={1 / 3}>1/3</Unit>
-      <Unit size={1 / 3}>1/3</Unit>
+      <Grid.Unit size={1 / 3}>
+        <Content>1/3</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 3}>
+        <Content>1/3</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 3}>
+        <Content>1/3</Content>
+      </Grid.Unit>
     </Grid>
 
     <Grid>
-      <Unit size={1 / 4}>1/4</Unit>
-      <Unit size={1 / 4}>1/4</Unit>
-      <Unit size={1 / 4}>1/4</Unit>
-      <Unit size={1 / 4}>1/4</Unit>
+      <Grid.Unit size={1 / 4}>
+        <Content>1/4</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 4}>
+        <Content>1/4</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 4}>
+        <Content>1/4</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 4}>
+        <Content>1/4</Content>
+      </Grid.Unit>
     </Grid>
 
     <Grid>
-      <Unit size={1 / 5}>1/5</Unit>
-      <Unit size={1 / 5}>1/5</Unit>
-      <Unit size={1 / 5}>1/5</Unit>
-      <Unit size={1 / 5}>1/5</Unit>
-      <Unit size={1 / 5}>1/5</Unit>
+      <Grid.Unit size={1 / 5}>
+        <Content>1/5</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 5}>
+        <Content>1/5</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 5}>
+        <Content>1/5</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 5}>
+        <Content>1/5</Content>
+      </Grid.Unit>
+      <Grid.Unit size={1 / 5}>
+        <Content>1/5</Content>
+      </Grid.Unit>
     </Grid>
   </>
 );