+++ packages/grid/src/tokens.ts
@@ -1,0 +1,3 @@
+import { BreakpointValueMap } from '@lendi-ui/breakpoint';
+
+export const gutter: BreakpointValueMap<string> = { xs: '0.4375rem', sm: '0.5rem', md: '0.625rem', lg: '0.75rem' };