+++ packages/grid/CHANGELOG.md
@@ -1,8 +1,11 @@
+## [4.0.0] 2020-07-01
+- Grid now implements its own gutters
+
 ## [3.0.1] 2019-09-26 [Release Notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
 - Bumped lendi-ui dependencies to latest versions
 
 ## [3.0.0] 2019-9-23 [Release notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
-### Updated 
+### Updated
 - Update typescript to version 3.5.2 and react to version 16.8.4
 
 ## [2.1.0] - 2019-05-17 - [HUB-174](https://creditandfinance.atlassian.net/browse/HUB-174)